Rename scripts/deploy-non-cf.sh
https://github.com/douglasnaphas/aws/issues/4

deploy-to-bucket.sh will be renamed deploy-to-frontend-bucket.sh. Then
I'll have deploy-to-non-cf-bucket.sh.

Basically separating the script that touches the bucket from the thing
that calls it.

Also the script that touches the non-cf bucket is going to deploy a bare
hello world-type index.html, which will be publicly readable, instead of
the actual frontend assets.