Deploy a bare site to non-cf, all envs
https://github.com/douglasnaphas/aws/issues/4

I have decided to merge this Athena/S3 data event logging experiment
into main. It's useful for the real template app to have S3 data event
logging.

I also still want a public bucket, so that I continue to illustrate the
answer to https://github.com/douglasnaphas/aws/issues/4. So this commit
makes the public bucket contents minimal, not the real brian-liberation
frontend code.

The bare hello-world HTML file is in a directory right alongside
deploy-to-non-cf-bucket.sh, to emphasize that it's not real application
code and is only for this non-CloudFront experiment.